completed directory refactor
