+++ Makefile
@@ -1,5 +1,5 @@
 # Default commit message
-COMMIT_MSG ?= "refactored test-lint job in CI"
+COMMIT_MSG ?= "completed directory refactor"
 
 # Target to add changes to staging
 add:
