+++ backend/v1/app/server/config.py
@@ -1,0 +1,5 @@
+""" Server Configurations for the Libooktrac Application"""
+
+TITLE="LiBookTrac App"
+VERSION="1.0.0"
+DESCRIPTION="library management system application backend"
