--- backend/v1/app/templates/index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>LiBookTrac Library Management Application</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            background: linear-gradient(to right, #8e44ad, #3498db);
-            color: #ffffff;
-            text-align: center;
-            margin: 0;
-            padding: 0;
-        }
-        h1 {
-            font-size: 3rem;
-            margin-top: 20vh;
-            text-shadow: 2px 2px 5px rgba(0, 0, 0, 0.3);
-        }
-        p {
-            font-size: 1.5rem;
-            margin-top: 1em;
-        }
-        footer {
-            position: fixed;
-            bottom: 10px;
-            width: 100%;
-            font-size: 0.9rem;
-            color: #dcdcdc;
-        }
-    </style>
-</head>
-<body>
-    <h1>Welcome to the LiBookTrac Application</h1>
-    <p>This application is under construction. We will be live soon.</p>
-    <footer>&copy; 2025 Abraham Ajibade. All rights reserved.</footer>
-</body>
-</html>
